ML Website
A Machine Learning project that consists of Banglore house price data, with which a model is developed using regression analysis. Given Parameters like Location, total square foot area, number of bathrooms and BHK (Bedroom, Hall, Kitchen / total number of rooms). A Flask website is also created on the same to obtain the output in a user friendly way.